[oh-tab-zhwi.2.3] Split extension activation tests by domain
